Add fill(), fix empty()
Decided to break apart the functionality of empty into a new method
called fill(). In the case where a user is not concerned if data is left
in the cbuffer until it is overwritten, then empty is the fastest
possible case. If they do need the data overwritten then they just call
fill() then empty().