Allow pop'd values to be gc'd
Changed .pop to remove the reference to the object. Also changed the
default value of .overflow to null. Changed how the default value of
rotateLeft/Right is detected.

The .pop method should remove the reference from the buffer, because
otherwise if it's a buffer of objects they'll never be garbage collected.
Since overflow is a function object, it makes sense that it starts out as
null. The methods rotateLeft and rotateRight should support receiving 0
(!cntr would evaluate to true and change it to 1 otherwise).